Renaming and moving of perf tests to align with functional counterparts.
