+++ tt_metal/tt-llk/tests/python_tests/perf_fast_tilize_full.py
@@ -34,7 +34,7 @@ from helpers.test_variant_parameters import (
     rt_dim=[1],
     ct_dim=[1, 2, 3, 4, 5, 6, 7, 8],
 )
-def test_perf_fast_tilize(perf_report, formats, rt_dim, ct_dim):
+def test_perf_fast_tilize_full(perf_report, formats, rt_dim, ct_dim):
     # Width 1 uses standard tilize fallback — not representative of fast path
     if ct_dim < 2:
         pytest.skip("ct_dim < 2 uses standard tilize fallback")
