+++ tt_metal/tt-llk/tests/python_tests/perf_sfpu_reduce.py
@@ -34,7 +34,7 @@ from helpers.test_variant_parameters import (
     reduce_pool=[ReducePool.Max],
     loop_factor=list(range(10, 201, 10)),
 )
-def test_perf_sfpu_reduce_row_max(
+def test_perf_sfpu_reduce(
     perf_report, formats, dest_acc, mathop, reduce_pool, loop_factor
 ):
     input_dimensions = [32, 32]
